@@ -3,6 +3,7 @@
   "version": "1.1.0",
   "description": "A Discord bot for use with your friends",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "build": "tsc --build",
     "clean": "tsc --build --clean",
@@ -17,24 +18,22 @@
   "author": "Racer159",
   "license": "ISC",
   "dependencies": {
-    "config": "^3.3.7",
-    "cowsay": "^1.4.0",
-    "discord.js": "^13.8.0",
+    "@types/giphy-api": "^2.0.3",
+    "@types/urban-dictionary": "^3.0.0",
+    "config": "^3.3.8",
+    "cowsay": "^1.5.0",
+    "discord.js": "^14.6.0",
     "fortune-teller": "^0.1.2",
-    "natural": "^5.2.2",
-    "request": "^2.88.2",
-    "request-promise": "^4.2.6",
-    "underscore": "^1.13.4"
+    "natural": "^5.2.3",
+    "node-fetch": "^3.2.10"
   },
   "devDependencies": {
-    "typescript": "^4.7.3",
-    "@types/node": "^18.0.0",
-    "@types/config": "0.0.38",
+    "@types/config": "3.3.0",
     "@types/natural": "^5.1.1",
-    "@types/request-promise": "^4.1.47",
-    "@types/underscore": "^1.11.4",
-    "@typescript-eslint/eslint-plugin": "^5.10.2",
-    "@typescript-eslint/parser": "^5.10.2",
-    "eslint": "^8.8.0"
+    "@types/node": "^18.11.0",
+    "@typescript-eslint/eslint-plugin": "^5.40.0",
+    "@typescript-eslint/parser": "^5.40.0",
+    "eslint": "^8.25.0",
+    "typescript": "^4.8.4"
   }
 }
