@@ -1,48 +1,54 @@
-import request = require('request-promise');
-import _ = require('underscore');
-import config = require('config');
+import fetch from 'node-fetch';
+import config from 'config';
+import * as errors from '../util/errors.js';
+import { MultiResponse, SingleResponse } from 'giphy-api';
+import * as utilities from '../util/utils.js';
 
 const GIPHYTOKEN = config.get<string>('GIPHYTOKEN');
 const RATELIMITRESPONSE = config.get<string>('RATELIMITRESPONSE');
 
 export async function gif(tokens: string[]): Promise<string> {
   try {
-    const response = JSON.parse(await request('http://api.giphy.com/v1/gifs/search?q=' + escape(tokens.join('+')) + '&api_key=' + GIPHYTOKEN));
+    const response = await fetch('https://api.giphy.com/v1/gifs/search?q=' + encodeURIComponent(tokens.join('+')) + '&api_key=' + GIPHYTOKEN);
 
-    if (response.data.length) {
-      const data = _.shuffle(response.data);
-      const id = data[0].id;
-      const imageUrl = 'http://media3.giphy.com/media/' + id + '/giphy.gif';
+    if (response.ok) {
+      const data = (await response.json()) as MultiResponse;
+      if (data.data.length) {
+        utilities.shuffle(data.data);
+        const id = data.data[0].id;
+        const imageUrl = 'http://media3.giphy.com/media/' + id + '/giphy.gif';
 
-      return imageUrl;
-    } else {
-      return 'Sorry, couldn\'t find anything!';
-    }
-  } catch (err: any) {
-    console.error(err);
-    if (err.statusCode === 429) {
+        return imageUrl;
+      } else {
+        return 'Sorry, couldn\'t find any gifs!';
+      }
+    } else if (response.status === 429) {
       return RATELIMITRESPONSE;
     } else {
-      return 'An error with the Giphy API occurred. ' + (err.error ? err.error : 'Check the logs for more information');
+      throw response;
     }
+  } catch (e) {
+    return errors.handleError('Giphy API', e);
   }
 }
 
 export async function translate(tokens: string[]): Promise<string> {
   try {
-    const response = JSON.parse(await request('http://api.giphy.com/v1/gifs/translate?s=' + escape(tokens.join('+')) + '&api_key=' + GIPHYTOKEN));
+    const response = await fetch('https://api.giphy.com/v1/gifs/translate?s=' + encodeURIComponent(tokens.join('+')) + '&api_key=' + GIPHYTOKEN);
 
-    if (response.data.images) {
-      return response.data.images.downsized_large.url;
-    } else {
-      return 'Sorry, couldn\'t find anything!';
-    }
-  } catch (err: any) {
-    console.error(err);
-    if (err.statusCode === 429) {
+    if (response.ok) {
+      const data = (await response.json()) as SingleResponse;
+      if (data.data.images) {
+        return data.data.images.downsized_large.url;
+      } else {
+        return 'Sorry, couldn\'t find anything!';
+      }
+    } else if (response.status === 429) {
       return RATELIMITRESPONSE;
     } else {
-      return 'An error with the Giphy API occurred. ' + (err.error ? err.error : 'Check the logs for more information');
+      throw response;
     }
+  } catch (e) {
+    return errors.handleError('Giphy API', e);
   }
 }
