@@ -1,22 +1,23 @@
-import request = require('request-promise');
-import config = require('config');
+import fetch from 'node-fetch';
+import config from 'config';
+import * as errors from '../util/errors.js';
+import { ChuckNorrisResponse } from '../types/response.js';
+
 const RATELIMITRESPONSE = config.get<string>('RATELIMITRESPONSE');
 
 export async function norris(): Promise<string>{
   try {
-    const response = JSON.parse(await request('https://api.chucknorris.io/jokes/random'));
-    if (response.value) {
-      return response.value;
-    } else {
-      return 'Sorry, there was an error recieving norris joke. No results found.';
-    }
+    const response = await fetch('https://api.chucknorris.io/jokes/random');
 
-  } catch (err: any) {
-    console.error(err);
-    if (err.statusCode === 429) {
+    if (response.ok) {
+      const data = (await response.json()) as ChuckNorrisResponse;
+      return data.value;
+    } else if (response.status === 429) {
       return RATELIMITRESPONSE;
     } else {
-      return 'An error with the Norris API occurred. ' + (err.error ? err.error : 'Check the logs for more information');
+      throw response;
     }
+  } catch (e) {
+    return errors.handleError('Chuck Norris API', e);
   }
 }
