@@ -1,28 +1,32 @@
-import request = require('request-promise');
-import _ = require('underscore');
-import config = require('config');
+import fetch from 'node-fetch';
+import config from 'config';
+import * as errors from '../util/errors.js';
+import { NutritionixResponse } from '../types/response.js';
 
 const NUTRITIONIXTOKEN = config.get<string>('NUTRITIONIXTOKEN');
 const RATELIMITRESPONSE = config.get<string>('RATELIMITRESPONSE');
 
 export async function calories(tokens: string[]): Promise<string> {
   try {
-    const response = JSON.parse(await request('https://api.nutritionix.com/v1_1/search/' + escape(tokens.join('+')) + '?results=0%3A1&cal_min=0&cal_max=5000&fields=item_name%2Cbrand_name%2Cnf_calories&appId=' + NUTRITIONIXTOKEN));
+    const response = await fetch('https://api.nutritionix.com/v1_1/search/' + encodeURIComponent(tokens.join('+')) +
+      '?results=0%3A1&cal_min=0&cal_max=5000&fields=item_name%2Cbrand_name%2Cnf_calories&appId=' + NUTRITIONIXTOKEN);
 
-    if (response.hits.length) {
-      const data = _.shuffle(response.hits);
-      const cals = data[0].fields.nf_calories;
+    if (response.ok) {
+      const data = (await response.json()) as NutritionixResponse;
 
-      return "A " + tokens.join(' ') + " is " + cals + " calories.";
+      if (data.hits.length) {
+        const cals = data.hits[0].fields.nf_calories;
+
+        return 'A ' + tokens.join(' ') + ' is ' + cals + ' calories.';
+      } else {
+        return 'Sorry, couldn\'t find anything!';
+      }
+    } else if (response.status === 429) {
+      return RATELIMITRESPONSE;
     } else {
-      return 'Sorry, couldn\'t find anything!';
+      throw response;
     }
-  } catch (err: any) {
-    console.error(err);
-    if (err.statusCode === 429) {
-        return RATELIMITRESPONSE;
-    } else {
-        return 'An error with the Nutritionix API occurred. ' + (err.error ? err.error : 'Check the logs for more information');
-    }
+  } catch (e) {
+    return errors.handleError('Nutritionix API', e);
   }
 }
